requirements.txt
asgiref==3.8.1
Django==4.2.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
iniconfig==2.0.0
packaging==24.0
pillow==10.3.0
pluggy==1.5.0
PyJWT==2.8.0
pytest==8.2.0
pytest-django==4.8.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.0
uritemplate==4.1.0
